test: rename dvc.testing.test_*.py so pytest doesn't handle them as test files
