+++ tests/func/test_data_cloud.py
@@ -9,7 +9,7 @@ import dvc_data
 from dvc.cli import main
 from dvc.external_repo import clean_repos
 from dvc.stage.exceptions import StageNotFound
-from dvc.testing.test_remote import (  # noqa, pylint: disable=unused-import
+from dvc.testing.remote_tests import (  # noqa, pylint: disable=unused-import
     TestRemote,
 )
 from dvc.utils.fs import remove
